Moved the schema file to the common folder
