+++ x-pack/plugins/siem/common/detection_engine/schemas/response/import_rules_schema.ts
@@ -7,7 +7,7 @@
 import * as t from 'io-ts';
 
 /* eslint-disable @typescript-eslint/camelcase */
-import { success, success_count } from './schemas';
+import { success, success_count } from '../common/schemas';
 import { errorSchema } from './error_schema';
 /* eslint-enable @typescript-eslint/camelcase */
